+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Possible loading content while permissions are loaded.
+
 ## [1.2.0] - 2023-05-18
 
 ### Added